#137 [client] Uncontrolled classification inputs
In the classification page, two text inputs were displaying a wrong
behaviour: When the user edits the input with the caret not being at the
end of the value, the caret moves to the end.

I tried encapsulating the problematic text inputs inside a new component
to monitor the full lifecycle. The input is not remounted. It sees to
much renderings, but does not loses the focus. This means that it is
React who does not know where to put that caret. I tried updating React
to the latest version, but it does not fix that issue

Action taken:
- The two problematic inputs are now uncontrolled
- A user must press the Enter key or click the submit button (the
magnifying glass next to the input) to see the modification become
effective